Create Neon Number(127)
Write a program to check given number is Neon Number or not and display messages Neon Number or Not Neon Number.
A neon number is a number where the sum of digits of square of the number is equal to the igven number. The task is to check and print neon numbers in a range.